+++ NEWS.md
@@ -1,5 +1,7 @@
-# sclet 0.0.2.001
+# sclet 0.0.2.002
 
++ Adding cell-cell communication module. (2024-10-06, Sat)
+    - `runCellChat()` wrapper function for running CellChat.
 + `gene_summary_table()` to add gene summary information to marker gene table (2024-10-02, Wed)
 
 # sclet 0.0.2